+++ apps/formio-server/server.js
@@ -35,14 +35,6 @@ module.exports = function(options) {
   // Use the given config.
   var config = options.config || require('./config');
 
-  // Add jslogger if configured.
-  var jslogger = null;
-  if (config.jslogger) {
-    jslogger = require('jslogger')({key: config.jslogger});
-  }
-
-  var Logger = require('./src/logger/index')(config);
-
   // Ensure that we create projects within the helper.
   app.hasProjects = true;
 
@@ -209,7 +201,9 @@ module.exports = function(options) {
 
   // Download a submission pdf.
   debug.startup('Attaching middleware: PDF Download');
+
   const downloadPDF = [
+    app.formio.formio.middleware.alias,
     require('./src/middleware/aliasToken')(app),
     app.formio.formio.middleware.tokenHandler,
     app.formio.formio.middleware.params,
@@ -217,7 +211,9 @@ module.exports = function(options) {
     require('./src/middleware/download')(app.formio)
   ];
 
+  app.get('/project/:projectId/:formAlias/submission/:submissionId/download', downloadPDF);
   app.get('/project/:projectId/form/:formId/submission/:submissionId/download', downloadPDF);
+  app.get('/project/:projectId/:formAlias/submission/:submissionId/download/:fileId', downloadPDF);
   app.get('/project/:projectId/form/:formId/submission/:submissionId/download/:fileId', downloadPDF);
 
   debug.startup('Attaching middleware: PDF Upload');
@@ -364,8 +360,15 @@ module.exports = function(options) {
     app.use('/project/:projectId/form/:formId/validate', require('./src/middleware/validateSubmission')(app.formio));
 
     // Mount the error logging middleware.
-    debug.startup('Attaching middleware: Logger');
-    app.use(Logger.middleware);
+    debug.startup('Attaching middleware: Error Handler');
+    app.use((err, req, res, next) => {
+      /* eslint-disable no-console */
+      console.log('Uncaught exception:');
+      console.log(err);
+      console.log(err.stack);
+      /* eslint-enable no-console */
+      res.status(400).send(typeof err === 'string' ? {message: err} : err);
+    });
 
     debug.startup('Attaching middleware: File Storage');
     app.storage = require('./src/storage/index.js')(app);
@@ -383,23 +386,11 @@ module.exports = function(options) {
 
   // Do some logging on uncaught exceptions in the application.
   process.on('uncaughtException', function(err) {
-    if (config.jslogger && jslogger) {
-      /* eslint-disable no-console */
-      console.log('Uncaught exception:');
-      console.log(err);
-      console.log(err.stack);
-      /* eslint-enable no-console */
-
-      jslogger.log({
-        message: err.stack || err.message,
-        fileName: err.fileName,
-        lineNumber: err.lineNumber
-      });
-    }
-
-    if (Logger.middleware) {
-      Logger.middleware(err, {});
-    }
+    /* eslint-disable no-console */
+    console.log('Uncaught exception:');
+    console.log(err);
+    console.log(err.stack);
+    /* eslint-enable no-console */
 
     // Give the loggers some time to log before exiting.
     setTimeout(function() {