+++ apps/formio-server/src/hooks/alter/email.js
@@ -52,7 +52,7 @@ module.exports = app => (mail, req, res, params, cb) => {
         email: mail.to,
         formId: form._id.toString(),
         projectId: form.project.toString(),
-        licenseKey: getLicenseKey(req.primaryProject),
+        licenseKey: getLicenseKey(req),
       });
     }
   };
@@ -201,31 +201,41 @@ module.exports = app => (mail, req, res, params, cb) => {
           });
       });
     }, () => {
-      ssoTokens.forEach((ssoToken) => {
-        if (ssoToken.submission) {
-          ssoToken.token = formioServer.formio.hook.alter('token', {
-            user: {
-              _id: ssoToken.submission._id.toString()
-            },
-            form: {
-              _id: ssoToken.submission.form.toString()
-            }
-          }, formObjs[ssoToken.submission.form.toString()]);
-          delete ssoToken.token.exp;
-          ssoToken.token = jwt.sign(ssoToken.token, formioServer.formio.config.jwt.secret, {
-            expiresIn: ssoToken.expireTime * 60
-          });
+      async.eachSeries(ssoTokens, (ssoToken, nextToken) => {
+        if (!ssoToken.submission) {
+          return nextToken();
         }
-      });
+        formioServer.formio.mongoose.models.session.create({
+          project: formObjs[ssoToken.submission.form.toString()].project,
+          form: ssoToken.submission.form,
+          submission: ssoToken.submission._id,
+          source: 'email',
+        })
+          .catch(nextToken)
+          .then((session) => {
+            ssoToken.token = formioServer.formio.hook.alter('token', {
+              user: {
+                _id: ssoToken.submission._id.toString()
+              },
+              form: {
+                _id: ssoToken.submission.form.toString()
+              }
+            }, formObjs[ssoToken.submission.form.toString()], {session});
+            delete ssoToken.token.exp;
+            ssoToken.token = jwt.sign(ssoToken.token, formioServer.formio.config.jwt.secret, {
+              expiresIn: ssoToken.expireTime * 60
+            });
+            nextToken();
+          });
+      }, () => {
+        // Replace the string token with the one generated here.
+        let index = 0;
+        mail.html = mail.html.replace(util.tokenRegex, () => {
+          return ssoTokens[index++].token || 'INVALID_TOKEN';
+        });
 
-      // Replace the string token with the one generated here.
-      let index = 0;
-      mail.html = mail.html.replace(util.tokenRegex, () => {
-        return ssoTokens[index++].token || 'INVALID_TOKEN';
+        return resolve(mail);
       });
-
-      // TO-DO: Generate the token for this user.
-      return resolve(mail);
     });
   });