+++ apps/formio-server/src/hooks/alter/models.js
@@ -12,7 +12,6 @@ module.exports = app => models => {
     },
     controller: {
       type: String,
-      default: ''
     },
     revisions: {
       type: String,
@@ -26,6 +25,19 @@ module.exports = app => models => {
       required: true,
       default: 0
     }
+  });
+
+  // Add a suggested indexes.
+  models.form.schema.index({
+    project: 1,
+    type: 1,
+    deleted: 1,
+    modified: -1
+  });
+  models.form.schema.index({
+    project: 1,
+    name: 1,
+    deleted: 1
   });
 
   models.submission.schema.add({
@@ -55,6 +67,23 @@ module.exports = app => models => {
       index: true,
       default: 'submitted'
     }
+  });
+
+  // Add a suggested indexes.
+  models.submission.schema.index({
+    form: 1,
+    project: 1,
+    deleted: 1
+  });
+  models.submission.schema.index({
+    project: 1,
+    deleted: 1
+  });
+  models.submission.schema.index({
+    form: 1,
+    project: 1,
+    'data.email': 1,
+    deleted: 1
   });
 
   // Add additional models.