+++ apps/formio-server/src/hooks/settings.js
@@ -51,7 +51,6 @@ module.exports = function(app) {
       init: require('./on/init')(app),
       formRequest: require('./on/formRequest')(app),
       validateEmail: require('./on/validateEmail')(app),
-      email: require('./on/email')(app)
     },
     alter: {
       formio: require('./alter/formio')(app),
@@ -120,11 +119,7 @@ module.exports = function(app) {
         const premium = [
           'webhook',
           'oauth',
-          'office365contact',
-          'office365calendar',
-          'hubspotContact',
           'googlesheet',
-          'jira',
           'ldap',
           'sqlconnector',
         ];
@@ -190,7 +185,7 @@ module.exports = function(app) {
           return true;
         }
         const premium = [
-          'webhook', 'oauth', 'office365contact', 'office365calendar', 'hubspotContact', 'googlesheet', 'jira', 'ldap'
+          'webhook', 'oauth', 'googlesheet', 'ldap'
         ];
 
         // If the action does not have a name, or is not flagged as being premium, ignore it.
@@ -206,16 +201,6 @@ module.exports = function(app) {
 
       emailTransports(transports, settings, req, cb) {
         settings = settings || {};
-        const office365 = settings.office365 || {};
-        if (office365.tenant && office365.clientId && office365.email && office365.cert && office365.thumbprint) {
-          transports.push(
-            {
-              transport: 'outlook',
-              title: 'Outlook'
-            }
-          );
-        }
-
         // Limit basic and independent
         if (req && req.primaryProject) {
           if (req.primaryProject.plan === 'basic') {
@@ -280,13 +265,49 @@ module.exports = function(app) {
        * @returns {Object}
        *   The modified token.
        */
-      token(token, form) {
-        token.origin = formioServer.formio.config.apiHost;
-        token.form.project = form.project;
-        token.project = {
-          _id: form.project
-        };
+      token(token, form, req) {
+        // See https://tools.ietf.org/html/rfc7519
+        token.iss = formioServer.formio.config.apiHost;
+        token.sub = token.user._id;
+        token.jti = req.session._id;
+
+        // form and project are now stored in session so no longer need to include it in the token.
+        delete token.form;
+        delete token.project;
+
         return token;
+      },
+
+      /**
+       * Token has just been decoded, make sure all the correct data is there.
+       *
+       * @param content
+       * @param req
+       * @param cb
+       */
+      tokenDecode(token, req, cb) {
+        if (!token.jti) {
+          return cb(null, token);
+        }
+        return formioServer.formio.mongoose.models.session.findById(token.jti)
+          .then((session) => {
+            if (!session) {
+              return cb(null, token);
+            }
+
+            req.session = session.toObject();
+
+            cb(null, {
+              ...token,
+              form: {
+                _id: req.session.form.toString(),
+                project: req.session.project.toString(),
+              },
+              project: {
+                _id: req.session.project.toString(),
+              },
+            });
+          });
       },
 
       /**
@@ -313,6 +334,85 @@ module.exports = function(app) {
 
           return cb();
         });
+      },
+
+      /**
+       * Called when the user is logged in.
+       *
+       * @param user
+       * @param req
+       * @param cb
+       * @returns {*}
+       */
+      login(user, req, cb) {
+        const sessionModel = formioServer.formio.mongoose.models.session;
+
+        sessionModel.create({
+          project: user.project._id,
+          form: user.form._id,
+          submission: user._id,
+          source: 'login',
+        })
+          .catch(cb)
+          .then((session) => {
+            req.session = session.toObject();
+            cb();
+          });
+      },
+
+      /**
+       * See if a token is valid.
+       *
+       * @param req
+       * @param decoded
+       * @param user
+       * @param cb
+       * @returns {*}
+       */
+      validateToken(req, decoded, user, cb) {
+        if (!decoded.jti) {
+          return cb(new Error('Missing session.'));
+        }
+
+        req.skipTokensValidation = true;
+        if (!req.session) {
+          return cb('Session not found.');
+        }
+        if (req.session.logout) {
+          return cb('Session no longer valid.');
+        }
+
+        const {
+          session: sessionConfig,
+        } = formioServer.formio.config;
+
+        const now = Date.now();
+
+        if (sessionConfig.expireTime !== '' && (now - req.session.created) > (sessionConfig.expireTime * 60 * 1000)) {
+          return cb('Session expired.');
+        }
+
+        return cb();
+      },
+
+      invalidateTokens(req, res, cb) {
+        const userId = req.params.submissionId;
+        if (!userId) {
+          return cb(new Error('No user found.'));
+        }
+
+        req.skipTokensInvalidation = true;
+
+        return formioServer.formio.mongoose.models.session.updateMany({
+          _id: {$ne: formioServer.formio.util.idToBson(req.session._id)},
+          submission: formioServer.formio.util.idToBson(userId),
+          logout: {$eq: null},
+        },
+        {
+          logout: Date.now(),
+        })
+          .catch(cb)
+          .then(() => cb());
       },
 
       isAdmin(isAdmin, req) {
@@ -747,13 +847,13 @@ module.exports = function(app) {
           }
           // No project but anonymous.
           else {
-            if (_url === '/spec.json') {
-              return true;
-            }
-
             // This req is unauthorized.
             return false;
           }
+        }
+
+        else if (req.projectId  && _url === `/project/${req.projectId}/spec.json`) {
+          return true;
         }
 
         else if (req.projectId && req.token && req.url === `/project/${req.projectId}/report`) {
@@ -883,6 +983,27 @@ module.exports = function(app) {
           });
         };
 
+        const createFormRevision = (item, doc, {_vid, tag}, done) => {
+          doc.set('_vid', _vid);
+          doc.save((err, result) => {
+            if (err) {
+              return done(err);
+            }
+
+            const body = Object.assign({}, item);
+            body._rid = result._id;
+            body._vid = result._vid;
+            body._vuser = 'system';
+            body._vnote = `Deploy version tag ${tag}`;
+            delete body._id;
+            delete body.__v;
+
+            formioServer.formio.mongoose.models.formrevision.create(body, () => {
+              done(null, item);
+            });
+          });
+        };
+
         alters.form = (item, template, done) => {
           item.project = template._id;
           this.formMachineName(item.machineName, item, (err, machineName) => {
@@ -910,25 +1031,54 @@ module.exports = function(app) {
                 return done(null, item);
               }
 
-              doc.set('_vid', parseInt(doc._vid) + 1);
-              doc.save((err, result) => {
-                if (err) {
-                  return done(err);
-                }
-
-                const body = Object.assign({}, item);
-                body._rid = result._id;
-                body._vid = result._vid;
-                body._vuser = 'system';
-                body._vnote = `Deploy version tag ${template.tag}`;
-                delete body._id;
-                delete body.__v;
-
-                formioServer.formio.mongoose.models.formrevision.create(body, () => {
-                  done(null, item);
-                });
-              });
+              createFormRevision(item, doc, {
+                _vid: parseInt(doc._vid) + 1,
+                tag: template.tag,
+              }, done);
             });
+          });
+        };
+
+        const getFormRevisions = (item, done) => {
+          formioServer.formio.resources.form.model.findOne({
+            machineName: item.machineName,
+            deleted: {$eq: null},
+            project: formioServer.formio.util.idToBson(item.project),
+          }).exec((err, form) => {
+            if (err) {
+              return done(err);
+            }
+
+            if (!form || form._vid || !form.revisions) {
+              return done(null, form);
+            }
+
+            formioServer.formio.resources.formrevision.model.findOne({
+              _rid: form._id,
+            }).exec((err, formrevision) => {
+              if (err) {
+                return done(err);
+              }
+
+              return done(null, form, formrevision);
+            });
+          });
+        };
+
+        alters.formSave = (item, done) => {
+          getFormRevisions(item, (err, form, formrevision) => {
+            if (err) {
+              return done(err);
+            }
+            // If there is no form or formrevision already exists then skip it
+            if (!form || form._vid || formrevision) {
+              return done(null, item);
+            }
+
+            createFormRevision(item, form, {
+              _vid: parseInt(form._vid) + 1,
+              tag: '',
+            }, done);
           });
         };