+++ apps/formio-server/src/middleware/projectUtilization.js
@@ -14,7 +14,7 @@ const NodeCache = require('node-cache');
 const cache = new NodeCache();
 
 // Cache response for 3 hours.
-const CACHE_TIME =  process.env.CACHE_TIME || 3 * 60 * 60 * 1000;
+const CACHE_TIME =  process.env.CACHE_TIME || 3 * 60 * 60;
 
 module.exports = (formio) => async (req, res, next) => {
   // If this isn't for a project, don't check.
@@ -25,11 +25,38 @@ module.exports = (formio) => async (req, res, next) => {
   const projectId = req.currentProject._id.toString();
 
   try {
-    if (cache.get(projectId) && _.has(req.primaryProject, 'settings.licenseKey')) {
-      return next();
-    }
-
     let licenseKey = getLicenseKey(req);
+
+    if (cache.get(projectId) && _.has(req.primaryProject, 'settings.licenseKey')) {
+      // We are going to evaluate these after going to the next middleware so not to slow down the request. If there is
+      // an issue we will clear the cache which will cause the issue to be caught on the next request.
+      // eslint-disable-next-line callback-return
+      next();
+      // Check every 5 minutes in case something changes.
+      if (cache.getTtl(projectId) - Date.now() < (CACHE_TIME - 300) * 1000) {
+        try {
+          const license = await getLicense(formio, licenseKey);
+          await utilization({
+            ...getProjectContext(req),
+            licenseKey,
+          });
+
+          // Everything is still good, update the cache ttl.
+          if (license.data.plan === req.primaryProject.plan) {
+            cache.ttl(projectId, CACHE_TIME);
+          }
+          else {
+            // Plan has changed. Re-evaluate on next request.
+            cache.del(projectId);
+          }
+        }
+        catch (err) {
+          // Utilization is now disabled. Re-evaluate on next request.
+          cache.del(projectId);
+        }
+      }
+      return;
+    }
 
     // Always allow access to the formio base project.
     if (req.primaryProject.name === 'formio') {
@@ -78,6 +105,11 @@ module.exports = (formio) => async (req, res, next) => {
       }
     }
 
+    // Always allow access to the project endpoint.
+    if (req.url === `/project/${req.projectId}`) {
+      return next();
+    }
+
     const result = await utilization({
       ...getProjectContext(req),
       licenseKey,
@@ -109,6 +141,6 @@ module.exports = (formio) => async (req, res, next) => {
   }
   catch (err) {
     cache.del(projectId);
-    return next(err);
+    res.status(400).send(err);
   }
 };