+++ apps/formio-server/src/plans/Commercial.js
@@ -8,8 +8,8 @@ module.exports = class Commercial extends Plan {
       submissionRequests: 2000000,
       emails: 10000,
       pdfDownloads: 1000,
-      pdfs: 50,
-      stages: 10,
+      pdfs: 25,
+      stages: 5,
       plan: 'commercial',
     };
   }