+++ apps/formio-server/src/resources/ProjectResource.js
@@ -133,6 +133,7 @@ module.exports = (router, formioServer) => {
   // Fix project plan (pull from actual license instead of stored DB value)
   formio.middleware.licenseUtilization = require('../middleware/licenseUtilization').middleware(formio);
   formio.middleware.licenseRemote = require('../middleware/licenseRemote').middleware(formio);
+  formio.middleware.licenseValid = require('../middleware/licenseValid')(formio);
 
   const hiddenFields = ['deleted', '__v', 'machineName', 'primary'];
   const resource = Resource(
@@ -159,6 +160,7 @@ module.exports = (router, formioServer) => {
     ],
     beforePost: [
       formio.middleware.filterMongooseExists({field: 'deleted', isNull: true}),
+      formio.middleware.licenseValid,
       require('../middleware/fetchTemplate'),
       formio.middleware.checkTenantProjectPlan,
       formio.middleware.projectDefaultPlan,
@@ -199,6 +201,7 @@ module.exports = (router, formioServer) => {
     ],
     beforePut: [
       formio.middleware.filterMongooseExists({field: 'deleted', isNull: true}),
+      formio.middleware.licenseValid,
       formio.middleware.licenseRemote,
       formio.middleware.licenseUtilization,
       function(req, res, next) {
@@ -251,6 +254,10 @@ module.exports = (router, formioServer) => {
           });
 
           req.body.access = _.uniqBy(req.body.access, 'type');
+        }
+
+        if (!req.body.name && req.body.name !== '') {
+          req.body.name = req.currentProject.name;
         }
 
         // Reset the machine name so that it will regenerate.
@@ -355,22 +362,6 @@ module.exports = (router, formioServer) => {
       return res.status(200).json({available: !project});
     });
   });
-
-  // Expose the atlassian oauth endpoints.
-  const atlassian = require('../actions/atlassian/util')(formioServer);
-  router.post(
-    '/project/:projectId/atlassian/oauth/authorize',
-    formio.middleware.tokenHandler,
-    formio.middleware.restrictProjectAccess({level: 'admin'}),
-    atlassian.authorizeOAuth,
-  );
-
-  router.post(
-    '/project/:projectId/atlassian/oauth/finalize',
-    formio.middleware.tokenHandler,
-    formio.middleware.restrictProjectAccess({level: 'admin'}),
-    atlassian.storeOAuthReply,
-  );
 
   // Expose the sql connector endpoint
   const sqlconnector = require('../actions/sqlconnector/util')(formioServer);